backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
motor==3.7.1
pymongo==4.15.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
transformers==4.56.1
torch==2.8.0
sentence-transformers==5.1.0
scikit-learn==1.7.2
scipy==1.16.2
pydantic==2.11.9
starlette==0.48.0
